Fix: The last day of a page might have been incomplete
Data on the "earliest date" of a page might be incomplete, since Africa's
Talking's truncates after 10000 results (the maximum page size) counting
backwards from the "end date".  This cannot be helped by paging through
the results, as `:page` appears to have no effect if `:count` is 10000.
Hence we need to discard this potentially incomplete data and
re-request a complete record for that date (and then continue paging by
day from that date on).

A complication is that the request parameters "start date" and "end date"
are treated by Africa's Talking as UTC, while the dates in the response CSV
appear to be in UTC+1.  Thus some time zone conversions are needed, too.